@@ -12,13 +12,13 @@ android {
             storePassword properties.getProperty('storePassword')
         }
     }
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion "27.0.3"
 
     defaultConfig {
         applicationId "com.asdamp.x_day"
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 28
         vectorDrawables.useSupportLibrary = true
         compileOptions {
             sourceCompatibility JavaVersion.VERSION_1_8
