@@ -2,13 +2,11 @@ package com.asdamp.utility;
 
 
 
-import android.app.Activity;
 import android.app.DatePickerDialog;
 import android.app.Dialog;
-import android.app.TimePickerDialog;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.text.format.DateFormat;
+import androidx.fragment.app.DialogFragment;
+
 import android.widget.DatePicker;
 
 public class DatePickerFragment extends DialogFragment
